% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{cleanAtlasExport}
\alias{cleanAtlasExport}
\title{Clean the Atlas export dataframe}
\usage{
cleanAtlasExport(x)
}
\arguments{
\item{x}{dataframe, usually the ouput of the import function 'importAtlasExport(x)'}
}
\value{
dataframe
}
\description{
Clean the Atlas data export dataframe before signal detection analysis
(see ?importAtlasExport() and online ECDC Atlas: http://atlas.ecdc.europa.eu/public/index.aspx) \cr
dataset <- cleanAtlasExport( importAtlasExport(x = 'ECDC_surveillance_data_Anthrax.csv') ) \cr
}
\details{
The function will: \cr
- Filter only on case based indicators i.e. 'Reported Cases" \cr
- Create four additional time variables to ease the analysis: \cr
TimeUnit ('Year', 'Month', 'Week'), \cr
TimeYear (xxxx), \cr
TimeMonth (xx) \cr
TimeWeek(xx) \cr
- Keep only variables of interest i.e. "HealthTopic", "Population", "Time", "RegionName", "NumValue" \cr
}
\seealso{
\code{\link{importAtlasExport}}
}
