% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{importAtlasExport}
\alias{importAtlasExport}
\title{Import ECDC Atlas export csv file}
\usage{
importAtlasExport(x)
}
\arguments{
\item{x}{file name of a csv file, export from the ECDC Atlas
(e.g. x = 'ECDC_surveillance_data_Anthrax.csv')}
}
\value{
dataframe
}
\description{
Import ECDC Atlas csv file that you can export
from the online ECDC Atlas: http://atlas.ecdc.europa.eu/public/index.aspx
e.g. ECDC_surveillance_data_Anthrax.csv \cr
dataset <- importAtlasExport(x = 'ECDC_surveillance_data_Anthrax.csv')
}
\details{
The function will interpret missing reports '-' as NA values
}
