% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{plotSD}
\alias{plotSD}
\title{Plot the Signal Detection time series}
\usage{
plotSD(x, input, subRegionName, x.sts, x.algo)
}
\arguments{
\item{x}{data.frame (default EpiSignalDetection::SignalData)}

\item{input}{list of parameters from the Signal Detection Application
(i.e. list(disease, country, indicator, stratification, unit, daterange, algo, testingperiod))}

\item{subRegionName}{region name label to use in the plot, if different than input$RegionName (optional)}

\item{x.sts}{sts object (optional)}

\item{x.algo}{algo object (optional)}
}
\value{
plot
}
\description{
Plot the Signal Detection time series including historical data, alarm detection period and alarms
}
\examples{
#-- Setting the parameters to run the report for
input <- list(
disease = "Salmonellosis",
country = "EU-EEA - complete series",
indicator = "Reported cases",
stratification = "Confirmed cases",
unit = "Month",
daterange = c("2010-01-01", "2016-12-31"),
algo = "FarringtonFlexible",
testingperiod = 5
)

#-- Plotting the signal detection output
plotSD(input = input)

}
\seealso{
\code{\link{SignalData}}
}
