% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEpiSDApp.R
\name{runEpiSDApp}
\alias{runEpiSDApp}
\title{Run the EpiSignalDectection Shiny application}
\usage{
runEpiSDApp()
}
\description{
Run the Shiny interactive application for signal detection analysis
based on ECDC Atlas export data.
See ECDC Atlas: \href{http://atlas.ecdc.europa.eu/public/index.aspx}{ECDC Surveillance Atlas of Infections Diseases}
}
