% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEpiSDReport.R
\name{runEpiSDReport}
\alias{runEpiSDReport}
\title{Run the EpiSignalDetection report (HTML markdown)}
\usage{
runEpiSDReport(input, stratified, outputfile)
}
\arguments{
\item{input}{list of parameters from the Signal Detection Application
(i.e. list(disease, country, indicator, stratification, unit, daterange, algo, testingperiod))
(see also default parameters in
\code{system.file("SignalDetectionReport_HTML", "SignalDetectionReport.Rmd", package = "EpiSignalDetection")})}

\item{stratified}{a logical value indicating whether the report
should be stratified by "Population" variable or not (default FALSE)}

\item{outputfile}{output file name (e.g. 'report.html')
(default value is a temporary folder - file.path(tempdir(), "SignalDectectionReport.html"))}
}
\value{
An HTML report
}
\description{
Function to render the markdown report of alarms in HTML format for ECDC Signal Detection Report
}
\examples{
#-- Running the report as a standalone function
#runEpiSDReport()    #Definition of each input parameter
                     #is done one by one through the R console

#---> OR

#-- First setting the parameters to run the report for
input <- list(
disease = "Salmonellosis",
country = "Portugal",
indicator = "Reported cases",
stratification = "Confirmed cases",
unit = "Month",
daterange = c("2011-01-01", "2016-12-31"),
algo = "FarringtonFlexible",
testingperiod = 6
)

#-- Second running the report based on the EpiSignalDetection::SignalData dataset
#-- and store it in a temporary folder
runEpiSDReport(input = input)

#-- Running the report based on the EpiSignalDetection::SignalData dataset
#-- and store the HTML output 'test.html' in the folder 'C:/R/'
# runEpiSDReport(input = input, outputfile = "C:/R/test.html")

#-- Running the report based on external data
input <- list(
file = list(datapath = "C:/Users/Downloads/ECDC_surveillance_data_Pertussis.csv"),
disease = "Pertussis",
country = "Greece",
indicator = "Reported cases",
stratification = "All cases",
unit = "Month",
daterange = c("2011-12-01", "2016-12-01"),
algo = "FarringtonFlexible",
testingperiod = 3
)
# runEpiSDReport(input = input, stratified = TRUE)
        # !!GLRNB algo does not allow stratified reports for now


}
\seealso{
\code{\link{SignalData}}
}
